refactor: ♻️ extraxt Discovery load methods to dedicated module
`Discovery` class is part of the API and these methods depend on the `Client` that internal.

Preparation for: #https://github.com/h2oai/cloud-discovery/issues/170
