+++ src/h2o_discovery/__init__.py
@@ -2,11 +2,12 @@ from typing import Optional
 
 from h2o_discovery import async_client
 from h2o_discovery import client
-from h2o_discovery import discovery
+from h2o_discovery import load
 from h2o_discovery import lookup
+from h2o_discovery import model
 from h2o_discovery._version import __version__  # noqa: F401
 
-Discovery = discovery.Discovery
+Discovery = model.Discovery
 
 
 def discover(
@@ -25,7 +26,7 @@ def discover(
     """
     uri = lookup.determine_uri(environment, discovery_address)
     cl = client.Client(uri)
-    return Discovery.load(cl)
+    return load.load_discovery(cl)
 
 
 async def discover_async(
@@ -44,4 +45,4 @@ async def discover_async(
     """
     uri = lookup.determine_uri(environment, discovery_address)
     cl = async_client.AsyncClient(uri)
-    return await Discovery.load_async(cl)
+    return await load.load_discovery_async(cl)
