+++ tests/test_load.py
@@ -3,7 +3,7 @@ from unittest import mock
 
 import pytest
 
-import h2o_discovery
+from h2o_discovery import load
 from h2o_discovery import model
 
 
@@ -28,36 +28,36 @@ ENVIRONMENT_DATA = model.Environment(
 )
 
 
-def test_discovery_environment(mock_client):
+def test_load_environment(mock_client):
     # Given
     mock_client.get_environment.return_value = ENVIRONMENT_DATA
 
     # When
-    discovery = h2o_discovery.Discovery.load(mock_client)
+    discovery = load.load_discovery(mock_client)
 
     # Then
     assert discovery.environment == ENVIRONMENT_DATA
 
 
 @pytest.mark.asyncio
-async def test_discovery_environment_async(mock_client):
+async def test_load_environment_async(mock_client):
     # Given
     mock_client.get_environment.return_value = asyncio.Future()
     mock_client.get_environment.return_value.set_result(ENVIRONMENT_DATA)
 
     # When
-    discovery = await h2o_discovery.Discovery.load_async(mock_client)
+    discovery = await load.load_discovery_async(mock_client)
 
     # Then
     assert discovery.environment == ENVIRONMENT_DATA
 
 
-def test_discovery_services(mock_client):
+def test_load_services(mock_client):
     # Given
     mock_client.list_services.return_value = [SERVICE_RECORD]
 
     # When
-    discovery = h2o_discovery.Discovery.load(mock_client)
+    discovery = load.load_discovery(mock_client)
 
     # Then
     assert discovery.services["test-service"] == SERVICE_RECORD
@@ -74,13 +74,13 @@ SERVICE_RECORD = model.Service(
 
 
 @pytest.mark.asyncio
-async def test_discovery_services_async(mock_client):
+async def test_load_services_async(mock_client):
     # Given
     mock_client.list_services.return_value = asyncio.Future()
     mock_client.list_services.return_value.set_result([SERVICE_RECORD])
 
     # When
-    discovery = await h2o_discovery.Discovery.load_async(mock_client)
+    discovery = await load.load_discovery_async(mock_client)
 
     # Then
     assert discovery.services["test-service"] == SERVICE_RECORD
@@ -93,25 +93,25 @@ CLIENT_RECORD = model.Client(
 )
 
 
-def test_discovery_clients(mock_client):
+def test_load_clients(mock_client):
     # Given
     mock_client.list_clients.return_value = [CLIENT_RECORD]
 
     # When
-    discovery = h2o_discovery.Discovery.load(mock_client)
+    discovery = load.load_discovery(mock_client)
 
     # Then
     assert discovery.clients["test-client"] == CLIENT_RECORD
 
 
 @pytest.mark.asyncio
-async def test_discovery_clients_async(mock_client):
+async def test_load_clients_async(mock_client):
     # Given
     mock_client.list_clients.return_value = asyncio.Future()
     mock_client.list_clients.return_value.set_result([CLIENT_RECORD])
 
     # When
-    discovery = await h2o_discovery.Discovery.load_async(mock_client)
+    discovery = await load.load_discovery_async(mock_client)
 
     # Then
     assert discovery.clients["test-client"] == CLIENT_RECORD
